* use --backup-interval cli parameter when creating default etcdBackupConfigs
* remove old cronjobs if new backup controller is enabled

+++ pkg/clusterdeletion/etcdbackupconfig.go
@@ -27,23 +27,25 @@ import (
 	controllerruntimeclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-func (d *Deletion) cleanupEtcdBackupConfigs(ctx context.Context, cluster *kubermaticv1.Cluster) error {
+func (d *Deletion) cleanupEtcdBackupConfigs(ctx context.Context, cluster *kubermaticv1.Cluster, etcdBackupRestoreController bool) error {
 	if !kuberneteshelper.HasFinalizer(cluster, kubermaticapiv1.EtcdBackupConfigCleanupFinalizer) {
 		return nil
 	}
 
-	backupConfigs := &kubermaticv1.EtcdBackupConfigList{}
-	if err := d.seedClient.List(ctx, backupConfigs, controllerruntimeclient.InNamespace(cluster.Status.NamespaceName)); err != nil {
-		return fmt.Errorf("failed to get EtcdBackupConfigs: %v", err)
-	}
-
-	if len(backupConfigs.Items) > 0 {
-		for _, backupConfig := range backupConfigs.Items {
-			if err := d.seedClient.Delete(ctx, &backupConfig); err != nil {
-				return fmt.Errorf("failed to delete EtcdBackupConfig %q: %v", backupConfig.Name, err)
-			}
+	if etcdBackupRestoreController {
+		backupConfigs := &kubermaticv1.EtcdBackupConfigList{}
+		if err := d.seedClient.List(ctx, backupConfigs, controllerruntimeclient.InNamespace(cluster.Status.NamespaceName)); err != nil {
+			return fmt.Errorf("failed to get EtcdBackupConfigs: %v", err)
 		}
-		return nil
+
+		if len(backupConfigs.Items) > 0 {
+			for _, backupConfig := range backupConfigs.Items {
+				if err := d.seedClient.Delete(ctx, &backupConfig); err != nil {
+					return fmt.Errorf("failed to delete EtcdBackupConfig %q: %v", backupConfig.Name, err)
+				}
+			}
+			return nil
+		}
 	}
 
 	oldCluster := cluster.DeepCopy()
